Refactor reminder letter services to allow subclassing
Or, more accurately, make a superclass so we can create variants for the AD letters.
